attempt composer packaging
