+++ LitCalTestServer.php
@@ -5,10 +5,6 @@
 // error_reporting(E_ALL);
 
 require_once 'vendor/autoload.php';
-require_once 'includes/enums/LitSchema.php';
-require_once 'includes/enums/ICSErrorLevel.php';
-require_once 'includes/LitTest.php';
-require_once 'includes/Health.php';
 
 use Ratchet\Server\IoServer;
 use Ratchet\Http\HttpServer;
