+++ src/enums/LitColor.php
