+++ src/enums/LitCommon.php
